Moving tally capability into header to be used in testing as well
